Attempt to just provide patterns and use in existing passes.
