+++ iree/compiler/Dialect/Flow/Conversion/TensorToFlow/CMakeLists.txt
@@ -10,42 +10,15 @@
 
 iree_add_all_subdirs()
 
-iree_tablegen_library(
-  NAME
-    PassesIncGen
-  TD_FILE
-    "Passes.td"
-  OUTS
-    -gen-pass-decls Passes.h.inc
-)
-
-iree_cc_library(
-  NAME
-    PassHeaders
-  HDRS
-    "PassDetail.h"
-    "Passes.h"
-    "Passes.h.inc"
-  DEPS
-    ::PassesIncGen
-    MLIRPass
-    MLIRTransforms
-  PUBLIC
-)
-
 iree_cc_library(
   NAME
     TensorToFlow
   HDRS
-    "Passes.h"
+    "ConvertTensorToFlow.h"
   SRCS
     "ConvertTensorToFlow.cpp"
-    "Passes.cpp"
   DEPS
-    ::PassHeaders
-    ::PassesIncGen
     MLIRIR
-    MLIRPass
     MLIRStandard
     MLIRTensor
     MLIRTransforms
