--- iree/compiler/Dialect/Flow/Conversion/TensorToFlow/Passes.cpp
@@ -1,18 +1,0 @@
-// Copyright 2021 The IREE Authors
-//
-// Licensed under the Apache License v2.0 with LLVM Exceptions.
-// See https://llvm.org/LICENSE.txt for license information.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-
-#include "iree/compiler/Dialect/Flow/Conversion/TensorToFlow/Passes.h"
-
-namespace mlir {
-namespace iree_compiler {
-
-namespace {
-#define GEN_PASS_REGISTRATION
-#include "iree/compiler/Dialect/Flow/Conversion/TensorToFlow/Passes.h.inc"  // IWYU pragma: export
-}  // namespace
-
-}  // namespace iree_compiler
-}  // namespace mlir
