+++ iree/compiler/InputConversion/Common/CMakeLists.txt
@@ -51,8 +51,8 @@ iree_cc_library(
     MLIRStandard
     MLIRTensor
     MLIRTransforms
-    iree::compiler::Dialect::Flow::Conversion::TensorToFlow
     iree::compiler::Dialect::Flow::IR
+    iree::compiler::Dialect::Flow::Transforms
   PUBLIC
 )
