+++ iree/compiler/InputConversion/Common/Passes.cpp
@@ -6,7 +6,7 @@
 
 #include "iree/compiler/InputConversion/Common/Passes.h"
 
-#include "iree/compiler/Dialect/Flow/Conversion/TensorToFlow/Passes.h"
+#include "iree/compiler/Dialect/Flow/Transforms/Passes.h"
 #include "mlir/Pass/PassManager.h"
 #include "mlir/Pass/PassOptions.h"
 #include "mlir/Pass/PassRegistry.h"
@@ -26,7 +26,9 @@ void registerCommonConversionPassPipelines() {
 
 // Common transformations to prepare input dialects for IREE.
 void buildCommonInputConversionPassPipeline(OpPassManager &passManager) {
-  passManager.addNestedPass<FuncOp>(IREE::Flow::createConvertTensorToFlow());
+  passManager.addNestedPass<FuncOp>(
+      IREE::Flow::createConvertToFlowTensorOpsPass(
+          /*runBeforeDispatchRegionFormation=*/true));
 }
 
 namespace {
